+++ CHANGELOG.md
@@ -14,6 +14,7 @@
  * Added `ElasticApmAttacher.attach(String propertiesLocation)` to specify a custom properties location
  * Logs message when `transaction_max_spans` has been exceeded (#849)
  * Report the number of affected rows by a SQL statement (UPDATE,DELETE,INSERT) in 'affected_rows' span attribute (#707)
+ * Add [`@Traced`](https://www.elastic.co/guide/en/apm/agent/java/master/public-api.html#api-traced) annotation which either creates a span or a transaction, depending on the context
 
 ## Bug Fixes
  * JMS creates polling transactions even when the API invocations return without a message